+++ assets/js/hooks/index.ts
@@ -180,6 +180,24 @@ function createKeyCombinationHook(
 }
 
 /**
+ * Function to dispatch a click event on the provided element.
+ *
+ * @param e - The keyboard event triggering the action.
+ * @param el - The HTML element to which the action will be applied.
+ */
+function clickAction(e: KeyboardEvent, el: HTMLElement) {
+  if (el.getAttribute('type') == 'submit') {
+    const formId = el.getAttribute('form');
+    const form = document.getElementById(formId);
+    form.dispatchEvent(
+      new Event('submit', { bubbles: true, cancelable: true })
+    );
+  } else {
+    el.dispatchEvent(new Event('click', { bubbles: true, cancelable: true }));
+  }
+}
+
+/**
  * Function to dispatch a submit event on the provided element.
  *
  * @param e - The keyboard event triggering the action.
@@ -201,8 +219,13 @@ function closeAction(e: KeyboardEvent, el: HTMLElement) {
 
 const isCtrlOrMetaS = (e: KeyboardEvent) =>
   (e.ctrlKey || e.metaKey) && e.key === 's';
+
 const isCtrlOrMetaEnter = (e: KeyboardEvent) =>
-  (e.ctrlKey || e.metaKey) && e.key === 'Enter';
+  (e.ctrlKey || e.metaKey) && !e.shiftKey && e.key === 'Enter';
+
+const isCtrlOrMetaShiftEnter = (e: KeyboardEvent) =>
+  (e.ctrlKey || e.metaKey) && e.shiftKey && e.key === 'Enter';
+
 const isEscape = (e: KeyboardEvent) => e.key === 'Escape';
 
 /**
@@ -216,9 +239,14 @@ export const SaveViaCtrlS = createKeyCombinationHook(
 /**
  * Hook to trigger a save and run action on the job panel when the Ctrl (or Cmd on Mac) + Enter key combination is pressed.
  */
-export const SaveAndRunViaCtrlEnter = createKeyCombinationHook(
+export const DefaultRunViaCtrlEnter = createKeyCombinationHook(
   isCtrlOrMetaEnter,
-  submitAction
+  clickAction
+);
+
+export const AltRunViaCtrlShiftEnter = createKeyCombinationHook(
+  isCtrlOrMetaShiftEnter,
+  clickAction
 );
 
 /**