+++ lib/lightning_web/live/workflow_live/manual_workorder.ex
@@ -24,7 +24,6 @@ defmodule LightningWeb.WorkflowLive.ManualWorkorder do
     <.form
       for={@form}
       id={@form.id}
-      phx-hook="SaveAndRunViaCtrlEnter"
       phx-change="manual_run_change"
       phx-submit="manual_run_submit"
       class="h-full flex flex-col gap-4"
@@ -35,7 +34,7 @@ defmodule LightningWeb.WorkflowLive.ManualWorkorder do
             type="select"
             field={@form[:dataclip_id]}
             options={@dataclips |> Enum.map(&{&1.id, &1.id})}
-            prompt="Create a new dataclip"
+            prompt="Create a new input"
             disabled={@disabled}
           />
         </div>
@@ -44,7 +43,7 @@ defmodule LightningWeb.WorkflowLive.ManualWorkorder do
       <div class="flex-0">
         <div class="flex flex-row">
           <div class="basis-1/2 font-semibold text-secondary-700 text-xs xl:text-base">
-            Dataclip Type
+            Type
           </div>
           <div class="basis-1/2 text-right">
             <Common.dataclip_type_pill type={
@@ -54,7 +53,7 @@ defmodule LightningWeb.WorkflowLive.ManualWorkorder do
         </div>
         <div class="flex flex-row mt-4">
           <div class="basis-1/2 font-semibold text-secondary-700 text-xs xl:text-base ">
-            State Assembly
+            Initial State
           </div>
           <div class="text-right text-xs xl:text-sm">
             <%= if(not is_nil(@selected_dataclip) and @selected_dataclip.type == :http_request) do %>